fixup! fixup! Integration tests fixes

+++ notary-eth-integration-test/src/integration-test/kotlin/integration/eth/DepositIntegrationTest.kt
@@ -130,7 +130,7 @@ class DepositIntegrationTest {
             integrationHelper.waitOneIrohaBlock()
 
             Assertions.assertEquals(
-                BigDecimal(amount, tokenInfo.precision.toInt()).add(BigDecimal(initialAmount)),
+                BigDecimal(amount, tokenInfo.precision).add(BigDecimal(initialAmount)),
                 BigDecimal(integrationHelper.getIrohaAccountBalance(clientIrohaAccountId, assetId))
             )
         }
@@ -166,7 +166,7 @@ class DepositIntegrationTest {
             integrationHelper.waitOneIrohaBlock()
 
             Assertions.assertEquals(
-                BigDecimal(amount, tokenInfo.precision.toInt()).add(BigDecimal(initialAmount)),
+                BigDecimal(amount, tokenInfo.precision).add(BigDecimal(initialAmount)),
                 BigDecimal(integrationHelper.getIrohaAccountBalance(clientIrohaAccountId, assetId))
             )
         }
